requirements.txt
altgraph==0.10.2
bdist-mpkg==0.5.0
beautifulsoup4==4.5.1
billboard.py==3.0.2
click==6.6
Flask==0.11.1
flask-heroku==0.1.9
Flask-SQLAlchemy==2.1
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
macholib==1.5.1
MarkupSafe==0.23
modulegraph==0.10.4
psycopg2==2.6.2
py2app==0.7.3
pyparsing==2.0.1
python-dateutil==1.5
pytz==2013.7
requests==2.11.1
six==1.4.1
sklearn==0.0
spotipy==2.3.8
SQLAlchemy==1.1.5
Werkzeug==0.11.11
xattr==0.6.4
zope.interface==4.1.1
